All working fine.

+++ src/main/java/be/cvo/broodservice/domain/BroodEntity.java
@@ -2,7 +2,6 @@ package be.cvo.broodservice.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Generated;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
@@ -17,6 +16,6 @@ public class BroodEntity {
     @Id
     private Integer id;
     private String name;
-    private Integer price;
+    private Double price;
     private String description;
 }
